Adding a case when there is only 1 system shortcut on the Popup Container
When there is only 1 system shortcut the Popup Container looks weird
because it shows the icon without text and it has a lot of space.

We already have a view to show the sorcuts with text and icons so I
reused that view and show it only if there is 1 system shortcut.

I also did a small refactor separating the sortcuts into
widgetShortcuts and nonWidgetShortcuts. The way I did it is not the
most efficient because it uses two passes, but I think it is easier to
read this way and we'll never have too many shortcuts.

Fix: 229356716
Test: Long press on the app icons in the workspace and in the
app drawer. Try different app some with one shortcut like the phone
app and some with multiple like.

Change-Id: Ie054af546758d0686914fc934772a76a3c690fe5