Cancel the recents animation for locked gesture[2/2]
When device is locked and going to finish the recents animation,
instead of start home activity, calling
finishRunningRecentsAnimation directly so core and shell does not need
to handle another transition.

Bug: 230582311
Test: enable shell transition. Start showWhenLocked activity while
keyguard showing. Scroll up to dismiss the activity and verify the
activity and home shall be stopped and only keyguard remains on the
screen.
Also verify entering recents works fine when device is unlocked.

Change-Id: Ic76f81efcaa8ed00d5af1475259ec16659f6b568