Update app icon shadows
Allow use of white shadow, set radius to 1.0 and offset to 0.
Use 100% and 20% alpha, instead of 70% and 50%.
(nets out to same effect)
Make ambient and key shadows equivalent: there are deterministic
formulas for what contrast a shadow will create, and those formulas
were used for the spec. Ad-hoc attempts to try to increase visibility
are no longer needed and don't align with design intent

Bug: 213314628
Test: Manual inspection at runtime
Change-Id: Ie12d3d05488e16fea94e91be16c25448f450a44d