Fix folder preview animation when adding or removing icons.
The functions PreviewItemManager.updatePreviewItems calls
PreviewItemManager.buildParamsForPage using mFirstPageParams and it
expects that mFirstPageParams contains the values before the reoredering
but they get updated in a previous call to updatePreviewItems(false),
so when doing the animation, the icons are animated to their current
position and nothing changes.

Fix: 233401274
Test: Add icons to a folder 1 to 5 without expanding the folder and
make sure they animate to their new position. Make sure they animate
when removing items as well.

Change-Id: Ia3e1393d8fb2b30eb107e8b2c6eefdf9e50e9f70