Allows work apps to be dragged down while work edu is showing.
Also fixes jumpiness of the personal/work tabs while scrolling
with this card showing.

Cause: AllAppsRecyclerView#getCurrentScrollY() calculates the
current scroll based on the top of the first item in the view.
This is used to determine if pulling down should move the
container (if you are scrolled to the top), and it is also used
to align the personal/work tabs as you scroll down (they stop
moving when the top item reaches their bottom).

Fix: The top of the Work Edu card was lower than the top of the
recycler view because it was offset by a top margin. Switching
this to padding in the card's container fixes the associated
calculations while keeping the visual positioning from before.

Fix: 233836148
Test: Manually tried dragging down and scrolling. Tried it with
and without my refactor.

Change-Id: Idaf4b3801174c1dfeceda5c0f112e405908e856d