Prevent wallpaper flicker while entering split screen from overview
When launching a pair of apps from overview panel, it will perform
cross-fade animations to transit from overview to the split pair. But
wallpaper might be visible for a while when both overveiw and split pair
is transparent during the transition. This updates to not fade-out
overview until the split pair is finishing showing in foregorund to
prevent the flicker of wallpaper.

Bug: 235137249
Test: swipe to home and enter overview panel, select and swipe down to
      enter a split pair, verified there's no flicker.
Change-Id: Iec1dc49e260f6e46640d47295d7f65dad5c24ab5