Replace shelf height with keep clear areas registration in Launcher.
This affects Hotseat only for now. Taskbar will be occluded when unstashed.

When in gesture nav and with auto-enter pip, we need to pre-register the Hotseat
keep clear areas, as otherwise the event appears after the animator is started
and current logic doesn't allow to update those destination bounds in the animator.

Test: manually, existing tests pass
Bug: 183746978

Change-Id: I4d97ca77225d3502acac1fb6b5e3eff3e81285ed