Keep the default display's top running task in the top of running list
While we move a task to virtual display, we will receive
onTaskMovedToFront() callback to add the task(on virtual display) into
the running list. It cause the first item of Recents is virtual
display's task but the thumbnail view is showing default display's
task snapshot.
This change is to keep default display's top task in the first while
adding a virtual display's task into the running task list.

Fix: 235797788
Test: 1.Launch app1 on default display,
        launch app2 on virtual display,
        enter Recents,
        observed the recent animation and overview is showing app1.
      2.Launch app1 on default display,
        swipe app1 to home,
        launch app2 on virtual display,
        enter Recents,
        observed the overview is showing app2.
https://drive.google.com/file/d/1XkGNupJknmFKqjE3c97s6ieCKN5v-n_z/view?resourcekey=0-NLDIaK2EV2tG2ItccUdxDw
Change-Id: Ibb740986782eadf02672081cbd716fed04be6468