Fixing findNearestArea to account for the padding.
The function findNearestArea calculates the corner of the
cell but in that calculation it doesn't account for the padding
in the cells.

Also, change ignoreOccupied because the name says one thing and
the code does the oposite so I change all true calls for false
calls and change the if to !ignoreOccupied so the description of
the parameter match the behaviour of the method.

Also, removing unused method.

Fix: 236134208
Fix: 236129504
Test: visualizing the components of findNearestArea in ag/19248816.
Change-Id: Iac50fbc76b4fa2acda21894ecb976ba612d468fe