Notifies All Apps header when profile tab switches.
This fixes scroll listeners which control the header from Taskbar.
In other instances of the All Apps container (i.e. from Launcher),
the search bar indirectly updates the header because resetSearch
is called, which also resets the header. Separately, I would like
to clean that up so that AllAppsContainerView does not get
controlled by the search bar (maybe as part of a larger reset
cleanup).

Fix: 236812759
Test: Manual on device with Taskbar
Change-Id: I9921e89c87d5e18e4b28854fbb373f24b4189014