Fix small bug with Overview action button placement
Overview actions will now be centered properly when 3-button nav is active on tablets in landscape.

On most tablet and foldable screens, we shift the Overview action buttons left to accommodate when 3-button nav is active. The only exception is large-screen tablets in landscape orientation, where there is enough room to center them. This patch adds a conditional for that case.

Fixes: 239247432
Test: Manual
Change-Id: I96f9d5e428273b0611463a1f0f5898910eefe5ce