Adding specific case for applications in getEntryMigrationId
A string like this are sued as id when migrating the grid:

The problem is that sourceBounds can chagne for the same app
and then return the same app as different. To fix it I added
a new case for Aplications and remove the variable sourceBounds
for the default case if there is anotherr Favorites item that
has the same problem but we don't know.

Test: Manually migrated between different grids.
Fix: 232004084
Change-Id: Ida015c971c70c8de2915fca1ebc38f446c21a310