Fix small bug with DWB banners not animating out promptly
This CL fixes an issue where initiating split select on a task with a DWB banner would result in the DWB banner sticking around long after the thumbnail itself fades, interfering visually with the rest of the animation.

Fixed by hiding all child views of the taskview when splitting, instead of just the thumbnail.

Fixes: 247257724
Test: Manual
Change-Id: I4dbeccf8dd28431749df5e78482e80aacf5df55a