Fix LauncherInstrumentation for OOP tests from gradle
The previous logic was assuming that if the test and target package were different, the target one was the launcher.

However, in the sysui-studio case, the packages are different despite none of the packages being launcher.

This detects if the test process is a gradle instrumentation, and if so selects the default launcher package name.

Test: locally with sysui-studio + presubmits
Bug: 234414088
Change-Id: Ie625e7d526f254f4b5f66b3a2b989d02eb95ac9c