Fix pause work apps overlapping IME
The bottom margin for button in gesture navigation is wrong.
The bottom margin for the button is supposed to be 20dp from the gesture nav bar,
not 20dp from the bottom of the screen. The calculation was correct for 3 button nav but
was forgotten to be applied to gesture nav.

bug: 246243359
test: Manual - after: https://hsv.googleplex.com/5763335673872384
Change-Id: If678a70b4ad43791cf9e66390164e1ab25a6e366