Replace drawRect with drawable for delightful pagination
In order to eventually use the shape referenced in the task (https://docs.google.com/presentation/d/1V9-1C3mJ3zRGsuzzzYExisr9vLYecgADa1FucMoW_aw/edit#slide=id.g14256d0c614_0_311) we need to use a drawable. This CL adds the drawable that will be changed to that shape and uses it in PageIndicatorDots.java instead of using drawRect.

Bug: 247561880
Test: manual
Change-Id: I01e10bd9ddc7e378f2dcfdee0f84ba7443c3b9d3