Prevent double touch inputs in the BaseDragLayer
If you hold a long press and while pressing you start another
long press again with another finger, then a new onLongPress event
will trigger and in some cases we are not processing that case.

By ignoring the evetns with the flag ACTION_OUTSIDE we can prevent
such cases.

Fix: 247477725
Test: ReorderWidgets
Change-Id: I31833bc54db2771809562bd1f92c8523eb2f05ab