Fix ViewCapture NPE.
The ViewPropertyRefs, when copied from 1 frame to another, where also
transferring the same "next" reference. Sometimes, the recycled
ViewPropertyRef would be the same ViewPropertyRef instance as one of the copied
ViewPropertyRefs, which would caused an NPE, since the recycled VPR
would have it's next value set to null.

Bug: 249125402
Test: Repro steps are dragging and dropping around an item on the
workspace to different locations. This consistently repro's the crash in
about 10 seconds of doing this. Afte the fix, the crash stopped.

Change-Id: Ia43d8f788eb4889c0890fea0bb839cb9fed78ada