Add transitions to DWB banner when app thumbnail progresses from Overview tile to fullscreen
This patch makes it so that the DWB banner transitions out neatly when a user in Overview drags a tile downward to launch the app. Previously, the banner would stick around throughout the transition, creating visual clutter and drawing attention to the corners of the app.

Fixed by creating an exit transition in setFullscreenProgress(), similar to other UI elements like Overview action chips.

There is another outstanding issue where sometimes the transition can get clipped and jump to the final frame. The cause is still unknown, but issue is tracked at b/250976138 for a future fix.

Fixes: 249825524
Test: Manual
Change-Id: Iea47a9bc643537f0b716ce11b104803d9ca25fd2