Change padding of widget recommendation table for correct size
Previously, the margin of the recommendation table was changed to fix the bug b/209579563 , but this leaves a bug where the left/right sides of the recommendation table would be too close to the screen border. This CL should leave enough space for both sides and reveal the scroll bar again, while keeping the talkback focused on the title first once it enters the widget picker.

Test: Manual
Bug: 206481237
Change-Id: I808c279c3264a51b21a0bb7af8ca455d284e36ff