Differentiate between app row and shortcut icons.
This is a simplified version of ag/20118087. It fixes the bug
with rows of shortcuts being in the search results (e.g. Gmail
and WhatsApp conversations), but it will not handle cases where
there are multiple rows of app icons. Currently I do not know of
any cases where that would happen.

Test: Manual with and without inject_web_top (which moves app row)
for WhatsApp and Gmail with AiAi fishfood.
Bug: 239927522

Change-Id: Ie9c62c0ee81086801942328c8770f37d5bf622c3