Fix bug with DWB banners causing Action Bar to vanish
This patch fixes a bug where entering Overview with 3+ DWB banners active would cause the Actions Bar to disappear, rendering it unusable.

The bug occurred because while the Actions Bar was animating itself in, the entry animation of the DWB banners caused a competing animation to be fired, which attempted to set the Actions Bar alpha back to 0 in a race-conditiony situation.

Fixed by using setAutoCancel() on the Actions View alpha animator, which auto-cancels a running animation when another animation is called on the same target, preventing race conditions.

Outstanding issue: 251277899
Fixes: 241163178
Test: Manual

Change-Id: I4edb5678b4aad5612cd5ef58141d8e22bf8587f0