Attempt to reduce flakiness of testDisableRotationCheckForPhone
The request setOrientationLeft() is async that puts the target
rotation to setting. So the actual display rotation change may
happen when executing goHome(). Then the swipe gesture may be
interrupted that causes the test to fail.

Add a condition to wait for leaving natural orientation change
before calling goHome(). Also reorder and simplify the way to
start test activity to reduce execution time.

Bug: 253290379
Test: TaplTestsQuickstep#testDisableRotationCheckForPhone
Change-Id: I881d6f666c98fd3fe601b874adb0d14122bff2b7