Set launcher state to NORMAL when entering desktop
When clicking on desktop tile, we need to explicitly move launcher state
to NORMAL. Freeform apps float on top of launcher and launcher remains
visible in the background. When on desktop, launcher should be in normal
state, with icons and widgets hidden. Hiding of the icons etc is handled
already by DesktopVisibilityController.

TODO:
 - need to fix home screen icons flicker, can be solved with new
   launcher state (b/255649902)

Bug: 244348395
Test: manual, click on desktop tile in overview, observe freeform apps
brought to front and overview closes

Change-Id: I214b0d6b622e673e0b0cb82f84b2080b479a7fb9