[QL-v2] Add logging for quick launch.
Move logging to DeviceSearchAdapterProvider to take advantage of View so we can log more info attached to the quick launch event. Another reason is currently we can't dynamically change the enter key icon so we can't really distinguish whether it's a quick search or quick launch using the icon (we are working with gboard team to fix this issue).

Bug: 256010819
Test: see wwlogging https://paste.googleplex.com/5668268562972672
Change-Id: I38aad2ae97c9b372f3d76222a82de45fc3bf3eed