Update timings for the Home <> All Apps transition.
Also updates some interpolators to make the manual and atomic
transitions more homogeneous.

Also also updates the commit point so it matches the transition point
between the two states exactly.

Note: the background scale is currently too much, and it needs to be 1/3
of the total distance to match the Workspace scaling. To do this, we
simply scale down the depth by a factor of 3 when calling the background
zoom-out API in the DepthController.

Bug: 254835725
Test: see videos in bug

Change-Id: I605546921e9ee4dfc852794c56495c72f0b916e1