Fix the wiggle reorder aniamtion stopping if dragging
We were using the minimum span to calculate the cell position
of the widget and if you move slightly because the minimum is usually smaller than the current size it would calculate wrong cell positions
that are change more frequently because it's smaller.

Also, I went back and moved
performReorder(... CellLayout.MODE_SHOW_REORDER_HINT) with the
alarm if block as it was before one of my refactors, it still works
if they are separate but is harder to think about the logic if they
are divided so it's better to all be in one block.

Fix: 255421416
Test: manual testing, reordering and making sure the animation is playing well.
Test: atest ReorderWidgets (to make sure the reorder is not broken)
Change-Id: I3eff5bc9f21fda8f2f8908380397d8639ac15ff9