Clear drag views during swipe up animation progress
This is to avoid clearing it too lately, at a point where the user starts to see a flicker on the screen.

There was already a report of this bug. Filing a bug report on my own because the original one is hard to find. I will mark that as fixed once I find it again.

Test: Manual
Fix: 259157574
Change-Id: Idc9bb67ebf6fc3ff6d78880f28c46d916a02ab77