Add onTaskMovedToFront to close floating views
Fixes a bug with allapps not being closed if the launching activity does
not trigger a onTaskStackChange callback. For example this is the case
when allapps launches a task in freeform mode. In this case the previous
task is still visible and won't be paused. The onTaskStackChange
callback is not triggered. Leaving the allapps view visible.

Bug: 259264346
Test: enable destop mode, open gmail, open allapps and open chrome,
  observe allapps view is closed
Change-Id: I5e60baa15ebc04b5094a84ed977eb9f3bc73d4a6